refactor: calculateResult 함수 분리
